+++ CHANGELOG.md
@@ -11,5 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - 카카오 로그인
 - 폰트 Pretandard 적용
+- 로그인 정보를 UserDefaults와 Keychain에 저장
 
 [unreleased]: https://github.com/Picplz/picplz-ios